+++ android-gif-drawable/src/main/jni/open_close.c
@@ -30,13 +30,13 @@ GifInfo *createGifHandle(GifSourceDescriptor *descriptor, JNIEnv *env, jboolean 
 		throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
 		return NULL;
 	}
-	info->controlBlock = calloc(sizeof(GraphicsControlBlock), 1);
-	info->controlBlock->DelayTime = DEFAULT_FRAME_DURATION_MS;
+	info->controlBlock = malloc(sizeof(GraphicsControlBlock));
 	if (info->controlBlock == NULL) {
 		DGifCloseFile(descriptor->GifFileIn);
 		throwException(env, OUT_OF_MEMORY_ERROR, OOME_MESSAGE);
 		return NULL;
 	}
+	setGCBDefaults(info->controlBlock);
 	info->destructor = NULL;
 	info->gifFilePtr = descriptor->GifFileIn;
 	info->startPos = descriptor->startPos;
@@ -52,7 +52,7 @@ GifInfo *createGifHandle(GifSourceDescriptor *descriptor, JNIEnv *env, jboolean 
 	info->backupPtr = NULL;
 	info->rewindFunction = descriptor->rewindFunc;
 	info->frameBufferDescriptor = NULL;
-	info->isOpaque = JNI_FALSE;
+	info->isOpaque = false;
 	info->sampleSize = 1;
 
 	DDGifSlurp(info, false, false);
@@ -94,4 +94,10 @@ GifInfo *createGifHandle(GifSourceDescriptor *descriptor, JNIEnv *env, jboolean 
 		return NULL;
 	}
 	return info;
+}
+
+void setGCBDefaults(GraphicsControlBlock *gcb) {
+	gcb->DelayTime = DEFAULT_FRAME_DURATION_MS;
+	gcb->TransparentColor = NO_TRANSPARENT_COLOR;
+	gcb->DisposalMode = DISPOSAL_UNSPECIFIED;
 }