+++ src/com/lee/cliplay/adapters/PreCachingLayoutManager.java
@@ -8,7 +8,7 @@ import android.support.v7.widget.RecyclerView;
  * Created by xl on 16/6/17.
  */
 public class PreCachingLayoutManager extends LinearLayoutManager {
-    private static final int DEFAULT_EXTRA_LAYOUT_SPACE = 600;
+    private static final int DEFAULT_EXTRA_LAYOUT_SPACE = 300;
     private int extraLayoutSpace = -1;
     private Context context;